% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{a_weibull}
\alias{a_weibull}
\title{A Weibull Raster for Austria}
\format{An object of class \code{RasterLayer} of dimension 2981 x 5811 x 1.}
\usage{
a_weibull
}
\description{
A Weibull Raster for Austria
}
\keyword{datasets}
