% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{BaroHoehe}
\alias{BaroHoehe}
\title{Deprecated use \code{\link[windfarmGA]{barometric_height}} instead.}
\usage{
BaroHoehe(data, height, po = 101325, ro = 1.225)
}
\arguments{
\item{data}{A data.frame containing the height values}

\item{height}{Column name of the height values}

\item{po}{Standardized air pressure at sea level (101325 Pa)}

\item{ro}{Standardized air density at sea level (1,225 kg per m3)}
}
\description{
BaroHoehe is replaced by
  \code{\link[windfarmGA]{barometric_height}}.
}
\keyword{internal}
