% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEn.R
\name{calculateEn}
\alias{calculateEn}
\title{Calculate Energy Outputs}
\usage{
calculateEn(sel, referenceHeight, RotorHeight, SurfaceRoughness, windraster,
  wnkl, distanz, polygon1, resol, RotorR, dirSpeed, srtm_crop, topograp,
  cclRaster)
}
\arguments{
\item{sel}{A list of the current population.}

\item{referenceHeight}{The height at which the incoming wind speeds
were measured. Default is 50m. (numeric)}

\item{RotorHeight}{The desired height of the turbine. Default is 100m.
(numeric)}

\item{SurfaceRoughness}{A surface roughness length of the
considered area in m. Default is 0.14 which is adequate for flat terrain.
(numeric)}

\item{windraster}{Calculated windraster for the considered area.
(raster)}

\item{wnkl}{Indicates the angle, at which no wake influences are
considered. Default is 20 degrees. (numeric)}

\item{distanz}{Indicates the distance, after which the wake effects are
considered to be eliminated. Default is 100km. (numeric)}

\item{polygon1}{The considered area as shapefile. (SpatialPolygons)}

\item{resol}{The resolution of the grid in meter. (numeric)}

\item{RotorR}{The desired rotor radius in meter. (numeric)}

\item{dirSpeed}{The wind speed and direction data.frame. (data.frame)}

\item{srtm_crop}{An SRTM raster for the considered area. Is only used,
when the terrain effect model is activated. (raster)}

\item{topograp}{Boolean value that indicates whether the
terrain effect model is activated ("TRUE") or deactivated ("FALSE").
(character)}

\item{cclRaster}{A Corine Land Cover raster, that has to be adapted
previously by hand with the surface roughness lenght for every land cover
type. Is only used, when the terrain effect model is activated.
(raster)}
}
\value{
Returns a list of all individuals of the current generation
with resulting wake effects, energy outputs, efficiency rates.
}
\description{
Calculate the energy output and efficiency rates of all
   individuals in the current population under all given wind directions
   and speeds. If the terrain effect model is activated, the main
   calculations to model the effects will be done in this function.
}
\author{
Sebastian Gatscha
}

