% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccl}
\alias{ccl}
\title{A Corine land Cover Raster}
\format{An object of class \code{RasterLayer} of dimension 46000 x 65000 x 1.}
\usage{
ccl
}
\description{
A Corine land Cover Raster
}
\references{
\url{https://www.eea.europa.eu/data-and-maps/data/clc-2006-raster-3/clc-2006-100m}
}
\keyword{datasets}
