% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapGA.R
\name{dup_coords}
\alias{dup_coords}
\title{Splits duplicated coords (copy of geoR)}
\usage{
dup_coords(x, ...)
}
\arguments{
\item{x}{Two column numeric matrix or data frame}

\item{...}{passed to sapply. If simplify = TRUE (default) results are
returned as an array if possible (when the number of replicates are the
same at each replicated location)}
}
\description{
This funtions takes an object with 2-D coordinates and returns
the positions of the duplicated coordinates. Also sets a method for
duplicated. Helper function for \code{\link{heatmapGA}}
}
\author{
Sebastian Gatscha
}
