% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRects.R
\name{getRects}
\alias{getRects}
\title{Get the Grid-IDs from binary matrix}
\usage{
getRects(trimtonOut, Grid)
}
\arguments{
\item{trimtonOut}{Input matrix with binary values. (matrix)}

\item{Grid}{Grid of the considered area (data.frame)}
}
\value{
Returns a list of all individuals with X and Y coordinates
and the grid cell ID. (list)
}
\description{
Get the grid IDs from the trimmed binary matrix, where the
binary code indicates which grid cells are used in the current
wind farm constellation.
}
\author{
Sebastian Gatscha
}

