% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoogleChromePlot.R
\name{GoogleChromePlot}
\alias{GoogleChromePlot}
\title{Plot the best Results in Google Chrome}
\usage{
GoogleChromePlot(result, Polygon1, best = 1, plotEn = 1, Projection)
}
\arguments{
\item{result}{The output matrix of \code{\link{genAlgo}}, which has
stored all relevant information. (matrix)}

\item{Polygon1}{The considered area as shapefile (SpatialPolygons)}

\item{best}{A numeric value indicating which best individuals
to be plotted. 1 will indicate the solution with highest value. (numeric)}

\item{plotEn}{A numeric value that indicates if the best energy or
efficiency output should be plotted. (plotEn==1) will plot the best
energy solution and (plotEn==2) will plot the best efficieny solution.
(numeric)}

\item{Projection}{A desired Projection can be used instead
of the default Lambert Azimuthal Equal Area Projection. (character)}
}
\description{
Plot the best energy or efficiency solutions in google Chrome
with a satelitte background image of google maps. Input Polygon is
not plotted
}
\author{
Sebastian Gatscha
}

