% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GooglePlot.R
\name{GooglePlot}
\alias{GooglePlot}
\title{Plot the the \code{best} Results with Google background map}
\usage{
GooglePlot(result, Polygon1, best = 1, plotEn = 1)
}
\arguments{
\item{result}{The output matrix of \code{\link{genAlgo}}, which has
stored all relevant information. (matrix)}

\item{Polygon1}{The considered area as shapefile (SpatialPolygons)}

\item{best}{A numeric value indicating which best individuals
to be plotted. 1 will indicate the solution with highest value. (numeric)}

\item{plotEn}{A numeric value that indicates if the best energy or
efficiency output should be plotted. (plotEn==1) will plot the best
energy solution and (plotEn==2) will plot the best efficieny solution.
(numeric)}
}
\value{
Returns a data.frame with the coordinates in LON/LAT and plots
the desired best result with a google background map. (data.frame)
}
\description{
Plot the best energy or efficiency solutions with a
background image of google maps.
}
\author{
Sebastian Gatscha
}

