% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GridFilter}
\alias{GridFilter}
\title{Deprecated use \code{\link[windfarmGA]{grid_area}} instead.}
\usage{
GridFilter(shape, resol = 500, prop = 1, plotGrid = FALSE)
}
\arguments{
\item{shape}{Simple Feature Polygon of the considered area}

\item{prop}{A factor used for grid calculation. Determines the percentage a
grid has to overlay the considered area to be represented as grid cell.
Default is 1.}

\item{plotGrid}{Logical value indicating whether the results should be
plotted. Default is FALSE.}
}
\description{
GridFilter is replaced by \code{\link[windfarmGA]{grid_area}}.
}
\keyword{internal}
