% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatMap}
\alias{heatMap}
\title{Plot heatmap of fit grid cells}
\usage{
heatMap(result, si = 2, Polygon1)
}
\arguments{
\item{result}{The output matrix of \code{\link{genAlgo}}, which
has stored all relevant informations (matrix)}

\item{si}{A numeric value that is used for the sizing of the
resolution of the heatmap. Default is 2 (numeric)}

\item{Polygon1}{The considered area as shapefile (SpatialPolygons)}
}
\description{
Plot a heatmap of the selected grid cells. Green grid
cells have been selected more often than red grid cells.
}
\author{
Sebastian Gatscha
}

