% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{HexaTex}
\alias{HexaTex}
\title{Deprecated use \code{\link[windfarmGA]{hexa_area}} instead.}
\usage{
HexaTex(Polygon1, size, plotTrue = FALSE)
}
\arguments{
\item{size}{The resolution of the grid in meters. Default is 500}
}
\description{
HexaTex is replaced by \code{\link[windfarmGA]{hexa_area}}.
}
\keyword{internal}
