% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hole_shape}
\alias{hole_shape}
\title{A shapefile with a hole}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 1 columns.
}
\usage{
hole_shape
}
\description{
A shapefile with a hole
}
\keyword{datasets}
