% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_population.R
\name{init_population}
\alias{init_population}
\title{Create a random initial Population}
\usage{
init_population(grid, n, n_start = 100)
}
\arguments{
\item{grid}{Indexed grid from \code{\link[=grid_area]{grid_area()}} (X, Y, cell IDs).}

\item{n}{A numeric value indicating the amount of required turbines.}

\item{n_start}{A numeric indicating the amount of randomly generated initial
individuals. Default is 100.}
}
\value{
Returns a list of \code{n_start} initial individuals, each consisting
of \code{n} turbines. Resulting list has the x and y coordinates, the grid
cell ID and a binary variable of 1, indicating a turbine in the grid cell.
}
\description{
Create \code{n_start} random sub-selections from the indexed
grid and assign binary variable 1 to selected grids. This function
initiates the genetic algorithm with a first random population and will
only be needed in the first iteration.
}
\examples{
library(sf)
## Exemplary input Polygon with 2km x 2km:
area <- sf::st_as_sf(sf::st_sfc(
  sf::st_polygon(list(cbind(
    c(4498482, 4498482, 4499991, 4499991, 4498482),
    c(2668272, 2669343, 2669343, 2668272, 2668272)
  ))),
  crs = 3035
))

Grid <- grid_area(area, 200, 1, TRUE)

## Create 5 individuals with 10 wind turbines each.
firstPop <- init_population(grid = Grid[[1]], n = 10, n_start = 5)

}
\seealso{
Other Genetic Algorithm Functions: 
\code{\link{crossover}()},
\code{\link{fitness}()},
\code{\link{genetic_algorithm}()},
\code{\link{mutation}()},
\code{\link{selection}()},
\code{\link{set_crossover}()},
\code{\link{swap_mutation}()},
\code{\link{trimton}()}
}
\concept{Genetic Algorithm Functions}
