% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{k_weibull}
\alias{k_weibull}
\title{K Weibull Raster for Austria}
\format{An object of class \code{RasterLayer} of dimension 2981 x 5811 x 1.}
\usage{
k_weibull
}
\description{
K Weibull Raster for Austria
}
\keyword{datasets}
