% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{leafPlot}
\alias{leafPlot}
\title{Deprecated use \code{\link[windfarmGA]{plot_leaflet}} instead.}
\usage{
leafPlot(result, Polygon1, which = 1, orderitems = TRUE, GridPol)
}
\arguments{
\item{result}{The resulting matrix of the function \code{\link{genetic_algorithm}}
or \code{\link{windfarmGA}}}

\item{Polygon1}{The considered area as SpatialPolygon, SimpleFeature Polygon
or coordinates as matrix/data.frame}

\item{which}{A numeric value, indicating which individual to plot. The
default is 1. Combined with \code{orderitems = TRUE} this will show the 
best performing wind farm.}

\item{orderitems}{A logical value indicating whether the results should be
ordered by energy values \code{TRUE} or chronologically \code{FALSE}}

\item{GridPol}{By default, the grid will be calculated based on the inputs 
of \code{result} and the \code{Polygon1}. But another spatial object or the 
output of the  \code{\link{grid_area}} or \code{\link{hexa_area}} functions 
can also be}
}
\description{
leafPlot is replaced by \code{\link[windfarmGA]{plot_leaflet}}.
}
\keyword{internal}
