% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multi_shape}
\alias{multi_shape}
\title{A MULTIPOLYGON with 3 Polygons}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 1 columns.
}
\usage{
multi_shape
}
\description{
A MULTIPOLYGON with 3 Polygons
}
\keyword{datasets}
