% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_random_search}
\alias{plot_random_search}
\title{Plot the result of a randomized output.}
\usage{
plot_random_search(resultRS, result, area, best)
}
\arguments{
\item{resultRS}{The result of the random functions
\code{\link{random_search_single}} and \code{\link{random_search}}.}

\item{result}{The output of \code{\link{genetic_algorithm}}}

\item{area}{Site polygon (\code{sf}, SpatialPolygons, or coordinate matrix).
Must be projected in metres.}

\item{best}{How many best candidates to plot. Default is 1.}
}
\value{
Returns NULL. Used for plotting
}
\description{
Plotting method for the results of
\code{\link{random_search_single}} and \code{\link{random_search}}.
}
\examples{
\donttest{
library(sf)
area <- sf::st_as_sf(sf::st_sfc(
  sf::st_polygon(list(cbind(
    c(4498482, 4498482, 4499991, 4499991, 4498482),
    c(2668272, 2669343, 2669343, 2668272, 2668272)
  ))),
  crs = 3035
))

Res <- random_search(result = resultrect, area = area)
plot_random_search(resultRS = Res, result = resultrect, area = area, best = 2)
}
}
\seealso{
Other Randomization: 
\code{\link{random_search}()},
\code{\link{random_search_single}()}
}
\concept{Randomization}
