% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbeorwor.R
\name{plotbeorwor}
\alias{plotbeorwor}
\title{Plot if previous population was better or worse}
\usage{
plotbeorwor(result)
}
\arguments{
\item{result}{The output matrix of \code{\link{windfarmGA}} or
\code{\link{genAlgo}}, which has stored all relevant information.}
}
\description{
Plot the changes in mean and max fitness values to previous
  generation.
}
\examples{
\donttest{
## Add some data examples from the package
load(file = system.file("extdata/resulthex.rda", package = "windfarmGA"))

## Plot the results of a hexagonal grid optimization
result <- resulthex
plotbeorwor(result)
}
}
\seealso{
Other Plotting Functions: \code{\link{PlotWindfarmGA}},
  \code{\link{RandomSearchTurb}}, \code{\link{heatmapGA}},
  \code{\link{interpol_view}}, \code{\link{plotCloud}},
  \code{\link{plotEvolution}}, \code{\link{plotResult}},
  \code{\link{plotWindrose}}, \code{\link{plot_viewshed}},
  \code{\link{plotfitnessevolution}},
  \code{\link{plotparkfitness}}
}
\concept{Plotting Functions}
