% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEvolution.R
\name{plotEvolution}
\alias{plotEvolution}
\title{Plot the evolution of fitness values}
\usage{
plotEvolution(result, ask = T, spar = 0.5)
}
\arguments{
\item{result}{The output matrix of \code{\link{genAlgo}}, which has
stored all relevant information. (matrix)}

\item{ask}{Should R wait for interaction for subsequent plotting.
Default is "T" (character)}

\item{spar}{A numeric value determining how exact a spline should
be drawn. Default is 0.5 (numeric)}
}
\description{
Plot the evolution of energy outputs and efficiency rates
over the whole generations. Plots min, mean and max values.
}
\author{
Sebastian Gatscha
}

