% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfitnessevolution.R
\name{plotfitnessevolution}
\alias{plotfitnessevolution}
\title{Plot the changes of min/mean/max fitness values}
\usage{
plotfitnessevolution(result, spar = 0.1)
}
\arguments{
\item{result}{An output matrix of the function \code{\link{windfarmGA}} or
\code{\link{genAlgo}} which has stored all relevant information.}

\item{spar}{A numeric value determining how exact a spline should be drawn.
Default is 0.1}
}
\description{
Plot the evolution of fitness values and the change in the min,
  mean and max fitness values to the former generations.
}
\examples{
\donttest{
## Add some data examples from the package
load(file = system.file("extdata/resulthex.rda", package = "windfarmGA"))

## Plot the results of a hexagonal grid optimization
result <- resulthex
plotfitnessevolution(result, 0.1)
}
}
\seealso{
Other Plotting Functions: \code{\link{PlotWindfarmGA}},
  \code{\link{RandomSearchTurb}}, \code{\link{heatmapGA}},
  \code{\link{interpol_view}}, \code{\link{plotCloud}},
  \code{\link{plotEvolution}}, \code{\link{plotResult}},
  \code{\link{plotWindrose}}, \code{\link{plot_viewshed}},
  \code{\link{plotbeorwor}}, \code{\link{plotparkfitness}}
}
\concept{Plotting Functions}
