% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotparkfitness.R
\name{plotparkfitness}
\alias{plotparkfitness}
\title{Plot the genetic algorithm results}
\usage{
plotparkfitness(result, spar = 0.5)
}
\arguments{
\item{result}{An output matrix of \code{\link{genAlgo}}, which has
stored all relevant information. (matrix)}

\item{spar}{A numeric value determining how exact a spline should
be drawn. Default is 0.1 (numeric)}
}
\description{
Plot the evolution of fitness values with the influences of
selection, crossover and mutation.
}
\author{
Sebastian Gatscha
}

