% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResults.R
\name{plotResult}
\alias{plotResult}
\title{Plot the best Results}
\usage{
plotResult(result, Polygon1, best = 5, plotEn = 1, topographie = FALSE,
  Grid, Projection)
}
\arguments{
\item{result}{An output matrix of the function \code{\link{genAlgo}},
which has stored all relevant information. (matrix)}

\item{Polygon1}{The considered area as shapefile. (SpatialPolygons)}

\item{best}{A numeric value indicating how many of the best individuals
should be plotted. (numeric)}

\item{plotEn}{A numeric value that indicates if the best energy or
efficiency output should be plotted. If (plotEn==1) plots the best energy
solutions and (plotEn==2) plots the best efficiency solutions. (numeric)}

\item{topographie}{A logical value, indicating whether terrain effects
should be considered and plotted or not. (logical)}

\item{Grid}{The grid as SpatialPolygons, which is obtained from
\code{\link{GridFilter}} and used for plotting.}

\item{Projection}{A desired Projection can be used instead
of the default Lambert Azimuthal Equal Area Projection. (character)}
}
\value{
Returns a data.frame of the best (energy/efficiency) individual
during all iterations. (data.frame)
}
\description{
Plot the best resulting solutions of the genetic algorithm.
Depending on \code{plotEn}, either the best energy or efficiency solutions
can be plotted. \code{best} indicates the amount of best solutions that
should be plotted.
}
\author{
Sebastian Gatscha
}

