% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWindfarmGA.R
\name{PlotWindfarmGA}
\alias{PlotWindfarmGA}
\title{Plot the results of an optimization run}
\usage{
PlotWindfarmGA(result, Polygon1, whichPl, best = 1, plotEn = 1, Projection)
}
\arguments{
\item{result}{An output matrix of the function \code{\link{windfarmGA}},
which has stored all relevant information. (matrix)}

\item{Polygon1}{The considered area as shapefile. Only required, if the
shapefile is already loaded.(SpatialPolygons)}

\item{whichPl}{Which Plots should be shown 1-8 are possible. If "all" is
given, then all plots will be shown.}

\item{best}{A numeric value indicating how many of the best individuals
should be plotted. (numeric)}

\item{plotEn}{A numeric value that indicates if the best energy or
efficiency output should be plotted. If (plotEn==1) plots the best energy
solutions and (plotEn==2) plots the best efficiency solutions. (numeric)}

\item{Projection}{A desired Projection can be used instead
of the default Lambert Azimuthal Equal Area Projection. (character)}
}
\description{
Plot the results of a genetic algorithm run with given inputs.
Several plots try to show all relevant influences and behaviours of the
algorithm. 8 plot methods are available that can be selected individually.
}
\author{
Sebastian Gatscha
}

