% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInteger.R
\name{readinteger}
\alias{readinteger}
\title{Check Input Crossover Method}
\usage{
readinteger()
}
\value{
Returns the selected crossover method (character)
}
\description{
Checks whether the input for \code{\link{crossover1}} is
given correctly. If not, a message is prompted which asks to input one
of the 2 available crossover methods. The available inputs are "E" and "R".
"E" refers to partitioning at equal intervals and "R" refers to random
partitioning.
}
\author{
Sebastian Gatscha
}

