% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resulthex}
\alias{resulthex}
\title{A resulting matrix of \code{genetic_algorithm} with 10 iterations
and a hexagonal grid derived from \code{sp_polygon}}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10 rows and 13 columns.
}
\usage{
resulthex
}
\description{
A resulting matrix of \code{genetic_algorithm} with 10 iterations
and a hexagonal grid derived from \code{sp_polygon}
}
\keyword{datasets}
