% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection1.R
\name{selection1}
\alias{selection1}
\title{Selection Method}
\usage{
selection1(fit, Grid, teil, elitism, nelit, selstate)
}
\arguments{
\item{fit}{A list of all fitness-evaluated individuals (list)}

\item{Grid}{Is the indexed grid output from \code{\link{GridFilter}}:
(data.frame)}

\item{teil}{A numeric value, that determines the selection percentage
(numeric)}

\item{elitism}{Boolean Value, which indicates whether elitism should be
included or not. (logical)}

\item{nelit}{If \code{elitism}: is TRUE, then this input variable
determines the amount of individuals in the elite group. (numeric)}

\item{selstate}{Determines which selection method is used, "FIX" selects
a constant percentage and "VAR" selects a variable percentage, depending
on the development of the fitness values. (character)}
}
\value{
Returns list with 2 elements. Element 1 is the binary encoded
matrix, which shows all selected individuals. Element 2 represent the mean
fitness values of each parental team. (list)
}
\description{
Select a certain amount of individuals and recombine them
to parental teams. Add the mean fitness value of both parents to
the parental team. Depending on the selected \code{selstate}, the
algorithm will either take always 50 percent or a variable percentage
of the current population. The variable percentage depends on the
evolution of the populations fitness values.
}
\author{
Sebastian Gatscha
}

