% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_polygon}
\alias{sp_polygon}
\title{The rectangular POLYGON used to create \code{resultrect} & \code{resulthex}}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 1 columns.
}
\usage{
sp_polygon
}
\description{
The rectangular POLYGON used to create \code{resultrect} & \code{resulthex}
}
\keyword{datasets}
