% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StartGA.R
\name{StartGA}
\alias{StartGA}
\title{Create a random initial Population}
\usage{
StartGA(Grid, n, nStart = 100)
}
\arguments{
\item{Grid}{The data.frame output of "GridFilter" function,
with X and Y coordinates and Grid cell IDs. (data.frame)}

\item{n}{A numeric value indicating the amount of required turbines.
(numeric)}

\item{nStart}{A numeric indicating the amount of randomly generated
initial individuals. Default is 100. (numeric)}
}
\value{
Returns a list of \code{nStart} initial individuals,
each consisting of \code{n} turbines.
Resulting list has the x and y coordinates, the grid cell ID
and a binary variable of 1, indicating a turbine in the grid cell.
(list)
}
\description{
Create \code{nStart} random subselections from the
indexed grid and assign binary variable 1 to selected grids.
This function initiates the genetic algorithm with a first random
population, and will only be needed in the firt iteration.
}
\author{
Sebastian Gatscha
}

