% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_model.R
\name{terrain_model}
\alias{terrain_model}
\title{Get terrainhic rasters}
\usage{
terrain_model(
  terrain = TRUE,
  area,
  ccl,
  ccl_roughness,
  plot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{terrain}{Terrain model (elevation + land cover).}

\item{area}{Site polygon (\code{sf}, SpatialPolygons, or coordinate matrix).
Must be projected in metres.}

\item{ccl}{Path to a Corine Land Cover raster when \code{terrain} is on.}

\item{ccl_roughness}{Path to the CLC legend CSV (\code{Rauhigkeit_z} column).}

\item{plot}{Plot the elevation and roughness rasters}

\item{verbose}{Print a line per generation.}
}
\value{
A list of SpatRasters
}
\description{
Calculate the SpatRasters needed for the terrain model.
}
\examples{
\dontrun{
library(sf)
area <- sf::st_as_sf(sf::st_sfc(
  sf::st_polygon(list(cbind(
    c(4651704, 4651704, 4654475, 4654475, 4651704),
    c(2692925, 2694746, 2694746, 2692925, 2692925)
  ))),
  crs = 3035
))
Polygon_wgs84 <- sf::st_transform(area, st_crs(4326))
srtm <- elevatr::get_elev_raster(locations = Polygon_wgs84, z = 11)
res <- terrain_model(srtm, area)
}
}
\concept{Terrain Model}
