% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexa_area.R
\name{tess2SPdf}
\alias{tess2SPdf}
\title{Create a Tesselation from a Polygon}
\usage{
tess2SPdf(x)
}
\arguments{
\item{x}{The Tesselation object (Tessellation)}
}
\value{
Returns a SpatialPolygons. (SpatialPolygons)
}
\description{
Returns a Spatial Polygons object from a Tesselation object.
}
\examples{
\dontrun{
library(spatstat)
library(maptools)
library(sp)
Polygon1 <- Polygon(rbind(c(4498482, 2668272), c(4498482, 2669343),
                          c(4499991, 2669343), c(4499991, 2668272)))
Polygon1 <- Polygons(list(Polygon1),1);
Polygon1 <- SpatialPolygons(list(Polygon1))
Projection <- "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000
+ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
proj4string(Polygon1) <- CRS(Projection)

## Create a hexagonal Tesselation
HexaGrid <- spatstat::hextess(maptools::as.owin.SpatialPolygons(Polygon1),s = 100)
plot(HexaGrid)

## Convert the Tesselation to SpatialPolygons
Hex2spdf <- tess2SPdf(HexaGrid)
plot(Hex2spdf)
}
}
\seealso{
Other Helper Functions: \code{\link{dup_coords}},
  \code{\link{getDEM}}, \code{\link{getISO3}},
  \code{\link{get_grids}}, \code{\link{grid_area}},
  \code{\link{hexa_area}}, \code{\link{isSpatial}},
  \code{\link{permutations}}, \code{\link{readintegerSel}},
  \code{\link{readinteger}}, \code{\link{splitAt}},
  \code{\link{windata_format}}
}
\concept{Helper Functions}
