% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visibility.R
\name{viewshed_par}
\alias{viewshed_par}
\title{Calculate visibility in parallel}
\usage{
viewshed_par(
  r,
  shape,
  turbine_locs,
  h1 = 0,
  h2 = 0,
  plot = FALSE,
  progress = "none",
  ...
)
}
\arguments{
\item{r}{A DEM raster}

\item{shape}{A Simpüle Feature Polygon of the windfarm area}

\item{turbine_locs}{Coordinates or SpatialPoint representing the wind
turbines}

\item{h1}{A numeric or numeric vector giving the extra height offsets for Point 1}

\item{h2}{A numeric giving the extra height offset for Point 2}

\item{plot}{Plot the process? Default is \code{FALSE}}

\item{...}{Additional arguments passed to the raster plot}
}
\description{
Calculate visibility for given points in a given area.
}
\seealso{
Other Viewshed Analysis: 
\code{\link{cansee}()},
\code{\link{interpol_view}()},
\code{\link{plot_viewshed}()},
\code{\link{rasterprofile}()},
\code{\link{viewTo}()},
\code{\link{viewshed}()}
}
\concept{Viewshed Analysis}
