% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{big_shape}
\alias{big_shape}
\title{A big shapefile}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 1 columns.
}
\usage{
big_shape
}
\description{
A big shapefile
}
\keyword{datasets}
