% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{calculateEn}
\alias{calculateEn}
\title{Deprecated use \code{\link[windfarmGA]{calculate_energy}} instead.}
\usage{
calculateEn(
  sel,
  referenceHeight,
  RotorHeight,
  SurfaceRoughness,
  wnkl,
  distanz,
  polygon1,
  resol,
  RotorR,
  dirSpeed,
  srtm_crop,
  topograp,
  cclRaster,
  weibull,
  plotit = FALSE
)
}
\arguments{
\item{sel}{A data.frame of an individual of the current population
(data.frame)}

\item{referenceHeight}{The height at which the incoming wind speeds were
measured (numeric)}

\item{RotorHeight}{The desired height of the turbines}

\item{SurfaceRoughness}{A surface roughness length of the considered area in
m. If the terrain effect model is activated, a surface roughness will be
calculated for every grid cell with the elevation and land cover
information}

\item{wnkl}{Indicates the angle at which no wake influences are considered
(numeric)}

\item{distanz}{Indicates the distance after which the wake effects are
considered to be eliminated}

\item{polygon1}{The considered area as shapefile}

\item{resol}{The resolution of the grid in meter}

\item{RotorR}{The desired rotor radius in meter}

\item{dirSpeed}{The wind speed and direction data.frame}

\item{srtm_crop}{A list of 3 raster, with 1) the elevation, 2) an orographic
and 3) a terrain raster. Calculated in \code{\link{genetic_algorithm}}}

\item{topograp}{Logical value that indicates whether the terrain effect model
is activated (TRUE) or deactivated (FALSE)}

\item{cclRaster}{A Corine Land Cover raster that has to be downloaded
previously. See also the details at \code{\link{windfarmGA}} The raster
will only be used when the terrain effect model is activated. (raster)}

\item{weibull}{A raster representing the estimated wind speeds}

\item{plotit}{Logical value. If TRUE, the process will be plotted. 
Default is FALSE.}
}
\description{
calculateEn is replaced by
  \code{\link[windfarmGA]{calculate_energy}}.
}
\keyword{internal}
