% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visibility.R
\name{cansee}
\alias{cansee}
\title{cansee}
\usage{
cansee(r, xy1, xy2, h1 = 0, h2 = 0)
}
\arguments{
\item{r}{A DEM raster}

\item{xy1}{A vector/matrix with X and Y coordinates for Point 1}

\item{xy2}{A vector/matrix with X and Y coordinates for Point 2}

\item{h1}{A numeric giving the extra height offset of Point 1}

\item{h2}{A numeric giving the extra height offset of Point 2}
}
\value{
A boolean value, indicating if the point (xy2) is visible
}
\description{
Check if point1 (xy1) visible from point2 (xy2) given 
a certain DEM (r)
}
