% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_energy.R
\name{circle_intersection}
\alias{circle_intersection}
\title{Get area of intersecting circles}
\usage{
circle_intersection(r1, r2, h1, h2, dx)
}
\arguments{
\item{r1}{The radius of circle 1}

\item{r2}{The radius of circle 2}

\item{h1}{The height of the circle center 1}

\item{h2}{The height of the circle center 2}

\item{dx}{The distance on the x-axis between both centers}
}
\value{
A numeric value
}
\description{
Calculate the intersection area of two circles with different
  radii and different heights
}
\seealso{
Other Wind Energy Calculation Functions: 
\code{\link{barometric_height}()},
\code{\link{calculate_energy}()},
\code{\link{get_dist_angles}()},
\code{\link{turbine_influences}()}
}
\concept{Wind Energy Calculation Functions}
