% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover1.R
\name{crossover1}
\alias{crossover1}
\title{Crossover Method}
\usage{
crossover1(se6, u, uplimit, crossPart)
}
\arguments{
\item{se6}{The selected individuals. The output of \code{\link{selection1}}
(list)}

\item{u}{The crossover point rate. (numeric)}

\item{uplimit}{The upper limit of allowed permutations. The current
algorithm has an upper bound of 300 permutations. (numeric)}

\item{crossPart}{The crossover method. Either "EQU" or "RAN". (character)}
}
\value{
Returns a binary coded matrix of all permutations and all grid
cells, 0 indicates no turbine and 1 indicates a turbine in the grid cell.
(matrix)
}
\description{
The crossover method of the genetic algorithm, which takes
the selected individuals after the \code{\link{selection1}} function and
produces new offsprings through permutation.
}
\author{
Sebastian Gatscha
}

