% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{crossover1}
\alias{crossover1}
\title{Deprecated use \code{\link[windfarmGA]{crossover}} instead.}
\usage{
crossover1(se6, u, uplimit, crossPart, verbose, seed)
}
\arguments{
\item{se6}{The selected individuals. The output of \code{\link{selection}}
(list)}

\item{u}{The crossover point rate. (numeric)}

\item{uplimit}{The upper limit of allowed permutations. The current algorithm
has an upper bound of 300 permutations. (numeric)}

\item{crossPart}{The crossover method. Either "EQU" or "RAN". (character)}

\item{verbose}{If TRUE, will print out further information.}

\item{seed}{Set a seed for comparability. Default is NULL}
}
\description{
crossover1 is replaced by \code{\link[windfarmGA]{crossover}}.
}
\keyword{internal}
