% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{dup_coords}
\alias{dup_coords}
\title{Splits duplicated coords (copy of geoR::dup.coords)}
\usage{
dup_coords(x, ...)
}
\arguments{
\item{x}{Two column numeric matrix or data frame}

\item{...}{passed to sapply. If simplify = TRUE (default) results are
returned as an array if possible (when the number of replicates are the
same at each replicated location)}
}
\value{
Function and methods returns NULL if there are no duplicates 
locations. Otherwise, the default method returns a list where each component
is a vector with the positions or the rownames, if available, of the
duplicates coordinates. The method for geodata returns a data-frame with
rownames equals to the positions of the duplicated coordinates, the first
column is a factor indicating duplicates and the remaining are output of
as.data.frame.geodata.
}
\description{
This function takes an object with 2-D coordinates and returns
the positions of the duplicated coordinates. Also sets a method for
duplicated. Helper function for \code{\link{plot_heatmap}}
}
\seealso{
Other Helper Functions: 
\code{\link{getDEM}()},
\code{\link{getISO3}()},
\code{\link{get_grids}()},
\code{\link{grid_area}()},
\code{\link{hexa_area}()},
\code{\link{isSpatial}()},
\code{\link{permutations}()},
\code{\link{readintegerSel}()},
\code{\link{readinteger}()},
\code{\link{splitAt}()},
\code{\link{tess2SPdf}()},
\code{\link{windata_format}()}
}
\author{
Paulo Justiniano Ribeiro Jr. \email{paulojus@leg.ufpr.br}
Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
\concept{Helper Functions}
