% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eucDist.R
\name{euc.dist}
\alias{euc.dist}
\title{Euclidian Distance between two Points}
\usage{
euc.dist(x, y)
}
\arguments{
\item{x}{A numeric value with X and Y coordinates for Point 1 (numeric)}

\item{y}{A numeric value with X and Y coordinates for Point 2 (numeric)}
}
\value{
Returns a numeric value indicating the euclidian distance between
two Points. (numeric)
}
\description{
Calculates the euclidian distance between two points.
}
\author{
Sebastian Gatscha
}

