% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{fitness}
\alias{fitness}
\title{Evaluate the Individual fFtness values}
\usage{
fitness(selection, referenceHeight, RotorHeight, SurfaceRoughness, Polygon,
  resol1, rot, dirspeed, srtm_crop, topograp, cclRaster)
}
\arguments{
\item{selection}{A list containing all individuals of the current
population. (list)}

\item{referenceHeight}{The height at which the incoming wind speeds were
measured. (numeric)}

\item{RotorHeight}{The desired height of the turbine. (numeric)}

\item{SurfaceRoughness}{A surface roughness length of the considered area
in m. (numeric)}

\item{Polygon}{The considered area as shapefile. (SpatialPolygons)}

\item{resol1}{The resolution of the grid in meter. (numeric)}

\item{rot}{The desired rotor radius in meter. (numeric)}

\item{dirspeed}{The wind speed and direction data.frame. (data.frame)}

\item{srtm_crop}{Logical value that indicates whether the terrain effect
model is activated (TRUE) or deactivated (FALSE). (logical)}

\item{topograp}{Logical value that indicates whether the terrain effect
model is activated (TRUE) or deactivated (FALSE). (logical)}

\item{cclRaster}{A Corine Land Cover raster, that has to be adapted
previously by hand with the surface roughness lenght for every land cover
type. Is only used, when the terrain effect model is activated. (raster)}
}
\value{
Returns a list with every individual, consisting of X & Y
coordinates, rotor radii, the runs and the selected grid cell IDs, and
the resulting energy outputs, efficiency rates and fitness values.
(list)
}
\description{
The fitness values of the individuals in the
current population are calculated, after having evaluated their energy
outputs in \code{\link{calculateEn}}. This function reduces the resulting
energy outputs to a single fitness value for every individual.
}
\author{
Sebastian Gatscha
}

