% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visibility.R
\name{getDEM}
\alias{getDEM}
\title{getDEM}
\usage{
getDEM(polygon, ISO3 = "AUT", clip = TRUE)
}
\arguments{
\item{polygon}{A Spatial / SimpleFeature Polygon to crop the DEM}

\item{ISO3}{The ISO3 code of the country}

\item{clip}{boolean, indicating if polygon should be cropped.
Default is TRUE}
}
\value{
A list with the DEM raster, and a SpatialPolygonsDataFrame or NULL
if no polygon is given
}
\description{
Get a DEM raster for a country based on ISO3 code
}
\examples{
\dontrun{
library(sp)
library(raster)
library(windfarmGA)
Polygon1 <- Polygon(rbind(c(4488182, 2667172), c(4488182, 2669343),
                          c(4499991, 2669343), c(4499991, 2667172)))
Polygon1 <- Polygons(list(Polygon1), 1);
Polygon1 <- SpatialPolygons(list(Polygon1))
Projection <- "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000
+ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
proj4string(Polygon1) <- CRS(Projection)
DEM_meter <- getDEM(Polygon1)
plot(DEM_meter[[1]])
plot(DEM_meter[[2]], add=T)
}
}
\author{
Sebastian Gatscha
}
