% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getRects}
\alias{getRects}
\title{Deprecated use \code{\link[windfarmGA]{get_grids}} instead.}
\usage{
getRects(trimtonOut, Grid)
}
\arguments{
\item{trimtonOut}{Input matrix with binary values.}

\item{Grid}{Grid of the considered area}
}
\description{
getRects is replaced by \code{\link[windfarmGA]{get_grids}}.
}
\keyword{internal}
