% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GooglePlot.R
\name{GooglePlot}
\alias{GooglePlot}
\title{Plot the \code{best} Results with Google background map}
\usage{
GooglePlot(result, Polygon1, best = 1, plotEn = 1, Projection, ...)
}
\arguments{
\item{result}{The output matrix of \code{\link{windfarmGA}} or
\code{\link{genAlgo}}, which has stored all relevant information.
(matrix)}

\item{Polygon1}{The considered area as shapefile (SpatialPolygons)}

\item{best}{A numeric value indicating which best individuals
to be plotted. 1 will indicate the solution with highest value. (numeric)}

\item{plotEn}{A numeric value that indicates if the best energy or
efficiency output should be plotted. (plotEn==1) will plot the best
energy solution and (plotEn==2) will plot the best efficiency solution.
(numeric)}

\item{Projection}{A desired Projection can be used instead
of the default Lambert Azimuthal Equal Area Projection. (character)}

\item{...}{If the plot is not required, it can be disabled with
the option \code{plT = FALSE}}
}
\value{
Returns a data.frame with the coordinates in LON/LAT and plots
the desired best result with a google background map. (data.frame)
}
\description{
Plot the best energy or efficiency solutions with a
background image of Google maps.
}
\examples{
\donttest{
## Add some data examples from the package
load(file = system.file("extdata/resultrect.rda", package = "windfarmGA"))
load(file = system.file("extdata/resulthex.rda", package = "windfarmGA"))
load(file = system.file("extdata/polygon.rda", package = "windfarmGA"))

## Plot the results of a wind farm optimization
result <- resultrect
Polygon1 <- polygon
GooglePlot(result, Polygon1, 1, 1)

}
}
\author{
Sebastian Gatscha
}
