% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{heatmapGA}
\alias{heatmapGA}
\title{Deprecated use \code{\link[windfarmGA]{plot_heatmap}} instead.}
\usage{
heatmapGA(result, si = 2, idistw)
}
\arguments{
\item{result}{The output of \code{\link{windfarmGA}} or
\code{\link{genetic_algorithm}}}

\item{si}{A numeric value that is used for the sizing of the resolution of
the heatmap. Default is 2}

\item{idistw}{The inverse distance weighting power. Default is the rotor
radius from the 'result' values}
}
\description{
heatmapGA is replaced by \code{\link[windfarmGA]{plot_heatmap}}.
}
\keyword{internal}
