% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{HexaTex}
\alias{HexaTex}
\title{Deprecated use \code{\link[windfarmGA]{hexa_area}} instead.}
\usage{
HexaTex(Polygon1, size, plotTrue = FALSE)
}
\arguments{
\item{Polygon1}{The SpatialPolygons object}

\item{size}{The side length of an hexagon}

\item{plotTrue}{Should the object be plotted}
}
\description{
HexaTex is replaced by \code{\link[windfarmGA]{hexa_area}}.
}
\keyword{internal}
