% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{InfluPoints}
\alias{InfluPoints}
\title{Deprecated use \code{\link[windfarmGA]{turbine_influences}} instead.}
\usage{
InfluPoints(t, wnkl, dist, polYgon, dirct, plotAngles = FALSE)
}
\arguments{
\item{t}{A data.frame of the current individual with X and Y coordinates}

\item{wnkl}{A numeric value indicating the angle, at which no wake influences
are considered. Default is 20 degrees.}

\item{dist}{A numeric value indicating the distance, after which the wake
effects are considered to be eliminated. Default is 100km.}

\item{polYgon}{A shapefile representing the considered area}

\item{dirct}{A numeric value indicating the current wind direction}

\item{plotAngles}{A logical variable, which is used to plot the distances and
angles. Default is FALSE.}
}
\description{
InfluPoints is replaced by
  \code{\link[windfarmGA]{turbine_influences}}.
}
\keyword{internal}
