% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genAlgo.R
\name{isSpatial}
\alias{isSpatial}
\title{Transform to SpatialPolygons}
\usage{
isSpatial(Polygon1, Projection)
}
\arguments{
\item{Polygon1}{An area as SpatialPolygon, SimpleFeature Polygon
or coordinates as matrix/data.frame}

\item{Projection}{Which Projection should be assigned to matrix / 
data.frame coordinates}
}
\value{
A SpatialPolygons object
}
\description{
Helper Function, which transforms SimpleFeatures or
coordinates as matrix/data.frame/data.table into a SpatialPolygon
}
\examples{
\donttest{
}
}
\author{
Sebastian Gatscha
}
