% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genAlgo.R
\name{isSpatial}
\alias{isSpatial}
\title{Transform to SpatialPolygons}
\usage{
isSpatial(Polygon1, Projection)
}
\arguments{
\item{Polygon1}{An area as SpatialPolygon, SimpleFeature Polygon
or coordinates as matrix/data.frame}

\item{Projection}{Which Projection should be assigned to matrix / 
data.frame coordinates}
}
\value{
A SpatialPolygons object
}
\description{
Helper Function, which transforms SimpleFeatures or
coordinates in matrix/data.frame/data.table into a SpatialPolygon
}
\details{
If the columns are named, it will look for common abbreviation
to match x/y or long/lat columns. If the columns are not named, the first
2 numeric columns are taken.
}
\examples{
\donttest{
df <- rbind(c(4498482, 2668272), c(4498482, 2669343),
            c(4499991, 2669343), c(4499991, 2668272))
isSpatial(df)

Polygon1 <- Polygon(rbind(c(4498482, 2668272), c(4498482, 2669343),
                          c(4499991, 2669343), c(4499991, 2668272)))
Polygon1 <- Polygons(list(Polygon1), 1);
Polygon1 <- SpatialPolygons(list(Polygon1))
Projection <- "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000
+ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
proj4string(Polygon1) <- CRS(Projection)
df_fort <- ggplot2::fortify(Polygon1)
isSpatial(df_fort)
}
}
\author{
Sebastian Gatscha
}
