% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafPlot.R
\name{leafPlot}
\alias{leafPlot}
\title{Leaflet Plot of a Wind Park}
\usage{
leafPlot(result, Polygon1, which = 1)
}
\arguments{
\item{result}{The resulting matrix of the function 'genAlgo' or
'windfarmGA'. (matrix)}

\item{Polygon1}{The Polygon for the wind farm area. (SpatialPolygons)}

\item{which}{A numeric value, indicating which best indidvual to plot.
The default is 1 (the best resulting wind farm). (numeric)}
}
\value{
Returns a leaflet map. (leaflet)
}
\description{
Plot a resulting wind farm on a leaflet map.
}
\examples{
\donttest{
library(sp)
## Create a random Polygon and plot it
Polygon1 <- Polygon(rbind(c(4651704, 2692925), c(4651704, 2694746),
                          c(4654475, 2694746), c(4654475, 2692925)))
Polygon1 <- Polygons(list(Polygon1),1);
Polygon1 <- SpatialPolygons(list(Polygon1))
Projection <- "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000
+ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
proj4string(Polygon1) <- CRS(Projection)
plot(Polygon1,axes=TRUE)

## Uniform wind speed and single wind direction
data.in <- as.data.frame(cbind(ws=12,wd=0))

## Run an optimization run with hexagonal grid cells.
result_hex <- genAlgo(Polygon1 = Polygon1, GridMethod ="h", n=12, Rotor=50,
                     fcrR=10,iteration=6, vdirspe = data.in,crossPart1 = "EQU",
                     selstate="FIX",mutr=0.8, Proportionality = 1,
                     SurfaceRoughness = 0.3, topograp = FALSE,
                     elitism=TRUE, nelit = 7, trimForce = TRUE,
                     referenceHeight = 50,RotorHeight = 100)

## Plot the resulting 'best' wind farm on a leaflet map.
leafPlot(result = result_hex, Polygon1 = Polygon1, which = 1)
}
}
\author{
Sebastian Gatscha
}
