% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{leafPlot}
\alias{leafPlot}
\title{Deprecated use \code{\link[windfarmGA]{plot_leaflet}} instead.}
\usage{
leafPlot(result, Polygon1, which = 1, orderitems = TRUE, GridPol)
}
\arguments{
\item{result}{The resulting matrix of the function \code{\link{genetic_algorithm}}
or \code{\link{windfarmGA}}}

\item{Polygon1}{The Polygon for the wind farm area.}

\item{which}{A numeric value, indicating which best individual to plot. The
default is 1 (the best resulting wind farm).}

\item{orderitems}{A logical value indicating whether the results should be
ordered by energy values (TRUE) or chronologically (FALSE).}

\item{GridPol}{The output grid polygon of the \code{\link{grid_area}} or
\code{\link{hexa_area}} functions.}
}
\description{
leafPlot is replaced by \code{\link[windfarmGA]{plot_leaflet}}.
}
\keyword{internal}
