% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multi_shape}
\alias{multi_shape}
\title{A multi-shapefile with 3 Polygons}
\format{
An object of class \code{SpatialPolygons} of length 1.
}
\usage{
multi_shape
}
\description{
A multi-shapefile with 3 Polygons
}
\keyword{datasets}
