% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{mutation}
\alias{mutation}
\title{Mutation Method}
\usage{
mutation(a, p)
}
\arguments{
\item{a}{The binary matrix of all individuals. (matrix)}

\item{p}{The mutation rate. (numeric)}
}
\value{
Returns a binary matrix with mutated genes. (matrix)
}
\description{
Mutate the genes of every chromosome or individual with
low probability.
}
\author{
Sebastian Gatscha
}

