% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_installed.R
\name{package_installed}
\alias{package_installed}
\alias{is_foreach_installed}
\alias{is_parallel_installed}
\alias{is_doparallel_installed}
\alias{is_ggplot2_installed}
\alias{is_leaflet_installed}
\alias{is_elevatr_installed}
\alias{is_plotly_installed}
\title{Is the package installed or not}
\usage{
is_foreach_installed()

is_parallel_installed()

is_doparallel_installed()

is_ggplot2_installed()

is_leaflet_installed()

is_elevatr_installed()

is_plotly_installed()
}
\value{
An invisible boolean value, indicating if the package is
installed or not.
}
\description{
Is the package installed or not
}
