% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cloud}
\alias{plot_cloud}
\title{Plot outputs of all generations with standard deviations}
\usage{
plot_cloud(result, pl = FALSE)
}
\arguments{
\item{result}{The output of \code{\link{genetic_algorithm}}}

\item{pl}{Should the results be plotted? Default is FALSE}
}
\value{
Returns a data.frame with the values for fitness, efficiency and
energy for all evaluated individuals
}
\description{
Plot the fitness, efficiency and energy outputs of all
generations and the corresponding standard deviations.
}
\examples{
\donttest{
## Plot the results of a hexagonal grid optimization
plcdf <- plot_cloud(resulthex, TRUE)
}
}
\seealso{
Other Plotting Functions: 
\code{\link{plot_development}()},
\code{\link{plot_evolution}()},
\code{\link{plot_fitness_evolution}()},
\code{\link{plot_parkfitness}()},
\code{\link{plot_result}()},
\code{\link{plot_windfarmGA}()},
\code{\link{plot_windrose}()},
\code{\link{random_search_single}()}
}
\concept{Plotting Functions}
