% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_development}
\alias{plot_development}
\title{Plot the progress of populations}
\usage{
plot_development(result)
}
\arguments{
\item{result}{The output of \code{\link{genetic_algorithm}}}
}
\value{
Returns NULL. Used for plotting
}
\description{
Plot the changes in mean and max fitness values to previous
generation.
}
\examples{
\donttest{
plot_development(resultrect)
}
}
\seealso{
Other Plotting Functions: 
\code{\link{plot_cloud}()},
\code{\link{plot_evolution}()},
\code{\link{plot_fitness_evolution}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_parkfitness}()},
\code{\link{plot_result}()},
\code{\link{plot_windfarmGA}()},
\code{\link{plot_windrose}()},
\code{\link{random_search_single}()}
}
\concept{Plotting Functions}
