% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_development.R
\name{plot_development}
\alias{plot_development}
\title{Plot the progress of populations}
\usage{
plot_development(result)
}
\arguments{
\item{result}{The output matrix of \code{\link{windfarmGA}} or
\code{\link{genetic_algorithm}}, which has stored all relevant information.}
}
\description{
Plot the changes in mean and max fitness values to previous
  generation.
}
\examples{
\donttest{
## Add some data examples from the package
load(file = system.file("extdata/resulthex.rda", package = "windfarmGA"))

## Plot the results of a hexagonal grid optimization
plot_development(resulthex)
}
}
\seealso{
Other Plotting Functions: \code{\link{interpol_view}},
  \code{\link{plot_cloud}}, \code{\link{plot_evolution}},
  \code{\link{plot_fitness_evolution}},
  \code{\link{plot_heatmap}},
  \code{\link{plot_parkfitness}},
  \code{\link{plot_result}}, \code{\link{plot_viewshed}},
  \code{\link{plot_windfarmGA}},
  \code{\link{plot_windrose}},
  \code{\link{random_search_single}}
}
\concept{Plotting Functions}
