% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visibility.R
\name{plot_farm_3d}
\alias{plot_farm_3d}
\title{Visibility Assessment}
\usage{
plot_farm_3d(DEM, turbines, zscale_label = 200, zscale_sphere = 200,
  zscale_3d = 200, txtu = "imhof1", txt = "Turbine", z = 1,
  sunangle = 315, ...)
}
\arguments{
\item{DEM}{The ISO3 code of the country}

\item{turbines}{boolean, indicating if polygon should be cropped.
Default is TRUE}

\item{z_scale}{A Spatial / SimpleFeature Polygon to crop the DEM}

\item{zoom1}{A Spatial / SimpleFeature Polygon to crop the DEM}

\item{windowsize}{A Spatial / SimpleFeature Polygon to crop the DEM}

\item{labl}{A Spatial / SimpleFeature Polygon to crop the DEM}
}
\value{

}
\description{
Plot a wind farm with rayshader in 3D
}
\examples{
\dontrun{
}
}
\author{
Sebastian Gatscha
}
