% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_fitness_evolution}
\alias{plot_fitness_evolution}
\title{Plot the changes of min/mean/max fitness values}
\usage{
plot_fitness_evolution(result, spar = 0.1)
}
\arguments{
\item{result}{The output of \code{\link{genetic_algorithm}}}

\item{spar}{A numeric value determining how exact a spline should be drawn.
Default is 0.1}
}
\value{
Returns NULL. Used for plotting
}
\description{
Plot the evolution of fitness values and the change in the min,
mean and max fitness values to the former generations.
}
\examples{
\donttest{
## Plot the results of a hexagonal grid optimization
plot_fitness_evolution(resulthex, 0.1)
}
}
\seealso{
Other Plotting Functions: 
\code{\link{plot_cloud}()},
\code{\link{plot_development}()},
\code{\link{plot_evolution}()},
\code{\link{plot_parkfitness}()},
\code{\link{plot_result}()},
\code{\link{plot_windfarmGA}()},
\code{\link{plot_windrose}()},
\code{\link{random_search_single}()}
}
\concept{Plotting Functions}
