% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitness_evolution.R
\name{plot_fitness_evolution}
\alias{plot_fitness_evolution}
\title{Plot the changes of min/mean/max fitness values}
\usage{
plot_fitness_evolution(result, spar = 0.1)
}
\arguments{
\item{result}{The output of function \code{\link{windfarmGA}} or
\code{\link{genetic_algorithm}}}

\item{spar}{A numeric value determining how exact a spline should be drawn.
Default is 0.1}
}
\description{
Plot the evolution of fitness values and the change in the min,
  mean and max fitness values to the former generations.
}
\examples{
\donttest{
## Add some data examples from the package
load(file = system.file("extdata/resulthex.rda", package = "windfarmGA"))

## Plot the results of a hexagonal grid optimization
plot_fitness_evolution(resulthex, 0.1)
}
}
\seealso{
Other Plotting Functions: 
\code{\link{interpol_view}()},
\code{\link{plot_cloud}()},
\code{\link{plot_development}()},
\code{\link{plot_evolution}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_parkfitness}()},
\code{\link{plot_result}()},
\code{\link{plot_viewshed}()},
\code{\link{plot_windfarmGA}()},
\code{\link{plot_windrose}()},
\code{\link{random_search_single}()}
}
\concept{Plotting Functions}
