% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_leaflet}
\alias{plot_leaflet}
\title{Plot a wind warm with leaflet}
\usage{
plot_leaflet(result, area, which = 1, orderitems = TRUE, grid = NULL)
}
\arguments{
\item{result}{The output of \code{\link{genetic_algorithm}}}

\item{area}{Site polygon (\code{sf}, SpatialPolygons, or coordinate matrix).
Must be projected in metres.}

\item{which}{A numeric value, indicating which individual to plot. The
default is 1. Combined with \code{orderitems = TRUE} this will show the
best performing wind farm.}

\item{orderitems}{A logical value indicating whether the results should be
ordered by energy values \code{TRUE} or chronologically \code{FALSE}}

\item{grid}{Optional grid polygons. By default they are rebuilt from
\code{result} and \code{area}. You can pass the polygon element of
\code{\link[=grid_area]{grid_area()}} or \code{\link[=hexa_area]{hexa_area()}}.}
}
\value{
Returns a leaflet map.
}
\description{
Plot a resulting wind farm on a leaflet map.
}
\examples{
\dontrun{
## Plot the best wind farm on a leaflet map (ordered by energy values)
plot_leaflet(result = resulthex, area = sp_polygon, which = 1)

## Plot the last wind farm (ordered by chronology).
plot_leaflet(
  result = resulthex, area = sp_polygon, orderitems = FALSE,
  which = 1
)

## Plot the best wind farm on a leaflet map with the rectangular Grid
Grid <- grid_area(sp_polygon, size = 150, prop = 0.4)
plot_leaflet(
  result = resultrect, area = sp_polygon, which = 1,
  grid = Grid[[2]]
)

## Plot the last wind farm with hexagonal Grid
Grid <- hexa_area(sp_polygon, size = 75)
plot_leaflet(
  result = resulthex, area = sp_polygon, which = 1,
  grid = Grid[[2]]
)
}
}
