% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbeorwor.R
\name{plotbeorwor}
\alias{plotbeorwor}
\title{Plot if previous population was better or worse}
\usage{
plotbeorwor(result)
}
\arguments{
\item{result}{The output matrix of \code{\link{genAlgo}}, which has
stored all relevant information. (matrix)}
}
\description{
Plot the changes in mean and max fitness values to previous
generation.
}
\author{
Sebastian Gatscha
}

