% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plotbeorwor}
\alias{plotbeorwor}
\title{Deprecated use \code{\link[windfarmGA]{plot_development}} instead.}
\usage{
plotbeorwor(result)
}
\arguments{
\item{result}{The output \code{\link{windfarmGA}} or 
\code{\link{genetic_algorithm}}}
}
\description{
plotbeorwor is replaced by
  \code{\link[windfarmGA]{plot_development}}.
}
\keyword{internal}
