% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcloud.R
\name{plotCloud}
\alias{plotCloud}
\title{Plot all outputs of all generations with coresponding
standard deviation}
\usage{
plotCloud(result, pl = "FALSE")
}
\arguments{
\item{result}{The output matrix of \code{\link{genAlgo}}, which has
stored all relevant information. (matrix)}

\item{pl}{Should the results be plotted? Default is "FALSE" (character)}
}
\value{
Returns a data.frame with the values for fitness, efficiency
and energy for all evaluated individuals. (data.frame)
}
\description{
Plot all the fitness, efficiency and energy outputs of
all generations. Plot the coresponding standard deviation below.
}
\author{
Sebastian Gatscha
}

