% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plotCloud}
\alias{plotCloud}
\title{Deprecated use \code{\link[windfarmGA]{plot_cloud}} instead.}
\usage{
plotCloud(result, pl = FALSE)
}
\arguments{
\item{result}{The output of \code{\link{windfarmGA}} or
\code{\link{genetic_algorithm}}}

\item{pl}{Should the results be plotted? Default is FALSE}
}
\description{
plotCloud is replaced by \code{\link[windfarmGA]{plot_cloud}}.
}
\keyword{internal}
