% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plotEvolution}
\alias{plotEvolution}
\title{Deprecated use \code{\link[windfarmGA]{plot_evolution}} instead.}
\usage{
plotEvolution(result, ask = TRUE, spar = 0.1)
}
\arguments{
\item{result}{The output of \code{\link{windfarmGA}} or
\code{\link{genetic_algorithm}}}

\item{ask}{Should R wait for interaction for subsequent plotting. Default is
TRUE}

\item{spar}{A numeric value determining how exact a spline should be drawn.
Default is 0.1}
}
\description{
plotEvolution is replaced by
  \code{\link[windfarmGA]{plot_evolution}}.
}
\keyword{internal}
