% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plotfitnessevolution}
\alias{plotfitnessevolution}
\title{Deprecated use \code{\link[windfarmGA]{plot_fitness_evolution}} instead.}
\usage{
plotfitnessevolution(result, spar = 0.1)
}
\arguments{
\item{result}{The output of function \code{\link{windfarmGA}} or
\code{\link{genetic_algorithm}}}

\item{spar}{A numeric value determining how exact a spline should be drawn.
Default is 0.1}
}
\description{
plotfitnessevolution is replaced by
  \code{\link[windfarmGA]{plot_fitness_evolution}}.
}
\keyword{internal}
