% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plotparkfitness}
\alias{plotparkfitness}
\title{Deprecated use \code{\link[windfarmGA]{plot_parkfitness}} instead.}
\usage{
plotparkfitness(result, spar = 0.1)
}
\arguments{
\item{result}{The output of \code{\link{windfarmGA}} or
\code{\link{genetic_algorithm}}}

\item{spar}{A numeric value determining how exact a spline should be drawn.
Default is 0.1}
}
\description{
plotparkfitness is replaced by
  \code{\link[windfarmGA]{plot_parkfitness}}.
}
\keyword{internal}
