% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plotResult}
\alias{plotResult}
\title{Deprecated use \code{\link[windfarmGA]{plot_result}} instead.}
\usage{
plotResult(
  result,
  Polygon1,
  best = 3,
  plotEn = 1,
  topographie = FALSE,
  Grid,
  Projection,
  sourceCCLRoughness,
  sourceCCL,
  weibullsrc
)
}
\arguments{
\item{result}{The output of \code{\link{windfarmGA}} or
\code{\link{genetic_algorithm}}}

\item{Polygon1}{The considered area as shapefile}

\item{best}{A numeric value indicating how many of the best individuals
should be plotted}

\item{plotEn}{A numeric value that indicates if the best energy or efficiency
output should be plotted. If (plotEn==1) plots the best energy solutions
and (plotEn==2) plots the best efficiency solutions}

\item{topographie}{A logical value, indicating whether terrain effects should
be considered and plotted or not}

\item{Grid}{The grid as SpatialPolygons, which is obtained from
\code{\link{grid_area}} and used for plotting}

\item{Projection}{A desired Projection can be used instead of the default
Lambert Azimuthal Equal Area Projection}

\item{sourceCCLRoughness}{The source to the adapted Corine Land Cover legend
as .csv file. Only required when terrain effect model is activated. As
default a .csv file within this package (\file{~/extdata/clc_legend.csv})
is taken that was already adapted manually}

\item{sourceCCL}{The source to the Corine Land Cover raster (.tif). Only
required, when the terrain effect model is activated}

\item{weibullsrc}{A list of Weibull parameter rasters, where the first list
item must be the shape parameter raster k and the second item must be the
scale parameter raster a of the Weibull distribution. If no list is given,
then rasters included in the package are used instead, which currently only
cover Austria.}
}
\description{
plotResult is replaced by
  \code{\link[windfarmGA]{plot_result}}.
}
\keyword{internal}
