% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{PlotWindfarmGA}
\alias{PlotWindfarmGA}
\title{Deprecated use \code{\link[windfarmGA]{plot_windfarmGA}} instead.}
\usage{
PlotWindfarmGA(
  result,
  Polygon1,
  GridMethod = "r",
  whichPl = "all",
  best = 1,
  plotEn = 1,
  Projection,
  weibullsrc
)
}
\arguments{
\item{result}{The output of \code{\link{windfarmGA}} or
\code{\link{genetic_algorithm}}}

\item{Polygon1}{The area as shapefile.}

\item{GridMethod}{Which grid spacing method was used. Default is
"rectangular". If hexagonal grid cells were used, assign any of the
following arguments: "h","hexa", "hexagonal"}

\item{whichPl}{Which plots should be shown: 1-6 are possible. The default is
"all" which shows all available plots}

\item{best}{A numeric value indicating how many of the best individuals
should be plotted}

\item{plotEn}{A numeric value that indicates if the best energy or efficiency
output is plotted. If (plotEn==1) plots the best energy solutions
and (plotEn==2) plots the best efficiency solutions}

\item{Projection}{A desired Projection can be used instead of the default
Lambert Azimuthal Equal Area Projection}

\item{weibullsrc}{A list of Weibull parameter rasters, where the first list
item must be the shape parameter raster `k` and the second item must be the
scale parameter raster `a` of the Weibull distribution. If no list is given,
then rasters included in the package are used instead, which currently only
cover Austria.}
}
\description{
PlotWindfarmGA is replaced by
  \code{\link[windfarmGA]{plot_windfarmGA}}.
}
\keyword{internal}
