% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plotWindrose}
\alias{plotWindrose}
\title{Deprecated use \code{\link[windfarmGA]{plot_windrose}} instead.}
\usage{
plotWindrose(
  data,
  spd,
  dir,
  spdres = 2,
  dirres = 10,
  spdmin = 1,
  spdmax = 30,
  palette = "YlGnBu",
  spdseq = NULL,
  plotit = TRUE
)
}
\arguments{
\item{data}{A data.frame containing the wind information}

\item{spd}{The column of the wind speeds in "data"}

\item{dir}{The column of the wind directions in "data"}

\item{spdres}{The increment of the wind speed legend. Default is 2}

\item{dirres}{The size of the wind sectors. Default is 10}

\item{spdmin}{Minimum wind speed. Default is 1}

\item{spdmax}{Maximal wind speed. Default is 30}

\item{palette}{A color palette used for drawing the wind rose}

\item{spdseq}{A wind speed sequence, that is used for plotting}

\item{plotit}{Should the windrose be plotted? Default is TRUE}
}
\description{
plotWindrose is replaced by
  \code{\link[windfarmGA]{plot_windrose}}.
}
\keyword{internal}
