% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{RandomSearch}
\alias{RandomSearch}
\title{Deprecated use \code{\link[windfarmGA]{random_search}} instead.}
\usage{
RandomSearch(result, Polygon1, n, best, Plot, GridMethod, max_dist = 2.2)
}
\arguments{
\item{result}{The resulting matrix of the function \code{\link{genetic_algorithm}}
or \code{\link{windfarmGA}}}

\item{Polygon1}{The Polygon for the wind farm area.}

\item{n}{The number of random searches to be performed. Default is 20.}

\item{best}{Which best individuals should be the starting conditions for a
random search. The default is 1.}

\item{Plot}{Should the random search be plotted? Default is FALSE}

\item{GridMethod}{Should the polygon be divided into rectangular or hexagonal
grid cells? The default is rectangular grid cells and hexagonal grid cells
are computed when assigning "h" or "hexagon" to this input variable. The
randomly generated points might be placed outside their hexagons.}

\item{max_dist}{A numeric value multiplied by the rotor radius to perform
collision checks. Default is 2.2}
}
\description{
RandomSearch is replaced by
  \code{\link[windfarmGA]{random_search}}.
}
\keyword{internal}
