% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomSearchPlot.R
\name{RandomSearchPlot}
\alias{RandomSearchPlot}
\title{RandomSearchPlot Plot the result of a randomized ouput.}
\usage{
RandomSearchPlot(resultRS, result, Polygon1, best)
}
\arguments{
\item{resultRS}{The result of the random functions \code{\link{RandomSearchTurb}}
and \code{\link{RandomSearch}}.}

\item{result}{The result of the function \code{\link{genAlgo}} or
\code{\link{windfarmGA}}.}

\item{Polygon1}{The Polygon for the wind farm area.}

\item{best}{How many best candidates to plot.}
}
\description{
Plotting method for the functions \code{\link{RandomSearchTurb}}
and \code{\link{RandomSearch}}.
}
\examples{
\donttest{
load(file = system.file("extdata/resultrect.rda", package = "windfarmGA"))
load(file = system.file("extdata/polygon.rda", package = "windfarmGA"))

Res = RandomSearchTurb(result = resultrect, Polygon1 = polygon, n=10)
RandomSearchPlot(resultRS = Res, result = resultrect, Polygon1 = polygon, best=2)
}
}
\author{
Sebastian Gatscha
}
