% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{RandomSearchPlot}
\alias{RandomSearchPlot}
\title{Deprecated use \code{\link[windfarmGA]{plot_random_search}} instead.}
\usage{
RandomSearchPlot(resultRS, result, Polygon1, best)
}
\arguments{
\item{resultRS}{The result of the random functions
\code{\link{random_search_single}} and \code{\link{random_search}}.}

\item{result}{The result of the function \code{\link{genetic_algorithm}} or
\code{\link{windfarmGA}}}

\item{Polygon1}{The Polygon for the wind farm area.}

\item{best}{How many best candidates to plot. Default is 1.}
}
\description{
RandomSearchPlot is replaced by
  \code{\link[windfarmGA]{plot_random_search}}.
}
\keyword{internal}
