% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{RandomSearchTurb}
\alias{RandomSearchTurb}
\title{Deprecated use \code{\link[windfarmGA]{random_search_single}} instead.}
\usage{
RandomSearchTurb(result, Polygon1, n, Plot, GridMethod, max_dist = 2.2)
}
\arguments{
\item{result}{The resulting matrix of the function \code{\link{genetic_algorithm}}
or \code{\link{windfarmGA}}}

\item{Polygon1}{The considered area as SpatialPolygon, SimpleFeature Polygon
or coordinates as matrix/data.frame}

\item{n}{The number of random searches to be performed. Default is 20.}

\item{Plot}{Should the random search be plotted? Default is \code{FALSE}}

\item{max_dist}{A numeric value multiplied by the rotor radius to perform
collision checks. Default is 2.2}
}
\description{
RandomSearchTurb is replaced by
  \code{\link[windfarmGA]{random_search_single}}.
}
\keyword{internal}
