% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resultrect}
\alias{resultrect}
\title{A resulting matrix of \code{genetic_algorithm} with 200 iterations 
and a rectangular shapefile \code{sp_polygon}}
\format{
An object of class \code{matrix} with 200 rows and 13 columns.
}
\usage{
resultrect
}
\description{
A resulting matrix of \code{genetic_algorithm} with 200 iterations 
and a rectangular shapefile \code{sp_polygon}
}
\keyword{datasets}
