% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{selection1}
\alias{selection1}
\title{Deprecated use \code{\link[windfarmGA]{selection}} instead.}
\usage{
selection1(fit, Grid, teil, elitism, nelit, selstate, verbose)
}
\arguments{
\item{fit}{A list of all fitness-evaluated individuals}

\item{Grid}{Is the indexed grid output from \code{\link{grid_area}}}

\item{teil}{A numeric value that determines the selection percentage}

\item{elitism}{Boolean value which indicates whether elitism should be
included or not.}

\item{nelit}{If \code{elitism} is TRUE, then this input variable determines
the amount of individuals in the elite group.}

\item{selstate}{Determines which selection method is used, "FIX" selects a
constant percentage and "VAR" selects a variable percentage, depending on
the development of the fitness values.}

\item{verbose}{If TRUE, will print out further information.}
}
\description{
selection1 is replaced by \code{\link[windfarmGA]{selection}}.
}
\keyword{internal}
