% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{splitAt}
\alias{splitAt}
\title{Divide matrices or integer at certain locations}
\usage{
splitAt(x, pos)
}
\arguments{
\item{x}{A numeric variable representing the binary genetic code of an
individual (numeric)}

\item{pos}{A numeric value which shows at which position the
genetic code is cut (numeric)}
}
\value{
Returns a list of the splitted genetic code.
}
\description{
Required function for the crossover method to
split a genetic code at random intervals. See also \code{\link{crossover}}.
}
\examples{
splitAt(1:100,20)
splitAt(as.matrix(1:100),20)

}
\seealso{
Other Helper Functions: 
\code{\link{dup_coords}()},
\code{\link{getDEM}()},
\code{\link{getISO3}()},
\code{\link{get_grids}()},
\code{\link{grid_area}()},
\code{\link{hexa_area}()},
\code{\link{isSpatial}()},
\code{\link{permutations}()},
\code{\link{readintegerSel}()},
\code{\link{readinteger}()},
\code{\link{tess2SPdf}()},
\code{\link{windata_format}()}
}
\concept{Helper Functions}
