% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{StartGA}
\alias{StartGA}
\title{Deprecated use \code{\link[windfarmGA]{init_population}} instead.}
\usage{
StartGA(Grid, n, nStart = 100)
}
\arguments{
\item{Grid}{The data.frame output of \code{\link{grid_area}}" function, with
X and Y coordinates and Grid cell IDs.}

\item{n}{A numeric value indicating the amount of required turbines.}

\item{nStart}{A numeric indicating the amount of randomly generated initial
individuals. Default is 100.}
}
\description{
StartGA is replaced by
  \code{\link[windfarmGA]{init_population}}.
}
\keyword{internal}
