% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain_model.R
\name{terrain_model}
\alias{terrain_model}
\title{Get topographic rasters}
\usage{
terrain_model(
  topograp = TRUE,
  Polygon1,
  sourceCCL,
  sourceCCLRoughness,
  plotit,
  verbose
)
}
\arguments{
\item{topograp}{Boolean value, which indicates if the terrain effect model
should be enabled or not. Default is \code{FALSE}}

\item{Polygon1}{The considered area as SpatialPolygon, SimpleFeature Polygon
or coordinates as matrix/data.frame}

\item{sourceCCL}{The path to the Corine Land Cover raster (.tif). Only
required when the terrain effect model is activated.}

\item{sourceCCLRoughness}{The source to the adapted Corine Land Cover legend
as .csv file. Only required when terrain effect model is activated. As
default a .csv file within this package (\file{~/extdata}) is taken that
was already adapted manually.}

\item{plotit}{If TRUE it will plot the best windfarm of every generation. 
Default is \code{FALSE}}

\item{verbose}{If TRUE it will print information for every generation.
Default is \code{FALSE}}
}
\value{
A list of SpatRasters
}
\description{
Calculate the SpatRasters needed for the terrain model.
}
\examples{
\donttest{
}
}
\concept{Terrain Model}
