% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexa_area.R
\name{tess2SPdf}
\alias{tess2SPdf}
\title{Create a Tesselation from a Polygon}
\usage{
tess2SPdf(x)
}
\arguments{
\item{x}{The Tesselation object (Tessellation)}
}
\value{
Returns a SpatialPolygons. (SpatialPolygons)
}
\description{
Returns a Spatial Polygons object from a Tesselation object.
}
\seealso{
Other Helper Functions: 
\code{\link{dup_coords}()},
\code{\link{getDEM}()},
\code{\link{getISO3}()},
\code{\link{get_grids}()},
\code{\link{grid_area}()},
\code{\link{hexa_area}()},
\code{\link{isSpatial}()},
\code{\link{permutations}()},
\code{\link{readintegerSel}()},
\code{\link{readinteger}()},
\code{\link{splitAt}()},
\code{\link{windata_format}()}
}
\concept{Helper Functions}
