% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimton.R
\name{trimton}
\alias{trimton}
\title{Adjust the amount of turbines per windfarm}
\usage{
trimton(mut, nturb, allparks, nGrids, trimForce)
}
\arguments{
\item{mut}{A binary matrix with the mutated individuals (matrix)}

\item{nturb}{A numeric value indicating the amount of required turbines
(numeric)}

\item{allparks}{A data.frame consisting of all individuals of the
current generation (data.frame)}

\item{nGrids}{A numeric value indicating the total amount of grid cells
(numeric)}

\item{trimForce}{A boolean value which determines which adjustment
method should be used. TRUE uses a probabilistic approach and
FALSE uses a random approach (logical)}
}
\value{
Returns a binary matrix with the correct amount of turbines
per individual (matrix)
}
\description{
Adjust the mutated individuals to the required amount of
turbines.
}
\author{
Sebastian Gatscha
}

