% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{VekWinkelCalc}
\alias{VekWinkelCalc}
\title{Deprecated use \code{\link[windfarmGA]{get_dist_angles}} instead.}
\usage{
VekWinkelCalc(t, o, wkl, distanz, polYgon, plotAngles)
}
\arguments{
\item{t}{A matrix of the current individual with x and y coordinates}

\item{o}{A numeric value indicating the index of the current turbine}

\item{wkl}{A numeric value indicating the angle, at which no wake influences
are considered. Default is 20 degrees.}

\item{distanz}{A numeric value indicating the distance, after which the wake
effects are considered to be eliminated. Default is 100km.}

\item{polYgon}{A shapefile representing the considered area}

\item{plotAngles}{A logical variable, which is used to plot the distances and
angles. Default is FALSE}
}
\description{
VekWinkelCalc is replaced by
  \code{\link[windfarmGA]{get_dist_angles}}.
}
\keyword{internal}
