% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visibility.R
\name{viewTo}
\alias{viewTo}
\title{viewTo}
\usage{
viewTo(r, xy1, xy2, h1 = 0, h2 = 0, progress = "none")
}
\arguments{
\item{r}{A DEM raster}

\item{xy1}{A matrix with X and Y coordinates for Point 1}

\item{xy2}{A matrix with X and Y coordinates for Points 2}

\item{h1}{A numeric giving the extra height offset of Point 1}

\item{h2}{A numeric giving the extra height offset of Point 2}

\item{progress}{Is passed on to plyr::aaply}
}
\value{
A boolean vector, indicating if Point 1 (xy1) is visible
from all elements of Points 2 (xy2)
}
\description{
Check if Point 1 (xy) is visible from multiple points
(xy2)
}
\author{
Sebastian Gatscha
}
