% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_algorithm.R
\name{windata_format}
\alias{windata_format}
\title{Transform Winddata}
\usage{
windata_format(df)
}
\arguments{
\item{df}{The wind data with speeds, direction and optionally a probability
column. If not assigned, it will be calculated}
}
\value{
A list of windspeed and probabilities
}
\description{
Helper Function, which transforms winddata to an acceptable
  format
}
\examples{
\donttest{
wind_df <- data.frame(ws = c(12, 30, 45), 
                      wd = c(0, 90, 150),
                      probab = 30:32)
windata_format(wind_df)

wind_df <- data.frame(speed = c(12, 30, 45), 
                      direction = c(90, 90, 150)
                      ,probab = c(10, 20, 60)
)
windata_format(wind_df)

wind_df <- data.frame(speed = c(12, 30, 45), 
                      direction = c(400, 90, 150)
)
windata_format(wind_df)
}
}
\seealso{
Other Helper Functions: 
\code{\link{dup_coords}()},
\code{\link{getDEM}()},
\code{\link{getISO3}()},
\code{\link{get_grids}()},
\code{\link{grid_area}()},
\code{\link{hexa_area}()},
\code{\link{isSpatial}()},
\code{\link{permutations}()},
\code{\link{readintegerSel}()},
\code{\link{readinteger}()},
\code{\link{splitAt}()},
\code{\link{tess2SPdf}()}
}
\concept{Helper Functions}
